+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- A task to list the translation progress, `crowdinListTranslationProgress`.
 
 ## [0.1.0] - 2021-07-26
 First alpha release.